Tests/LibWeb/Text/expected/link_as.txt

Link.as: style
Link.as: empty
